refactor: move yolox bbox conversion functions to new location
