+++ peekingduck/pipeline/nodes/model/yoloxv1/yolox_files/utils.py
@@ -136,31 +136,3 @@ def strip_optimizer(
         f"Original size: {orig_filesize}MB. "
         f"Stripped size: {stripped_filesize}MB."
     )
-
-
-def xywh2xyxy(inputs: torch.Tensor) -> torch.Tensor:
-    """Converts from [x, y, w, h] to [x1, y1, x2, y2] format.
-
-    (x, y) is the object center. (x1, y1) is the top left corner and (x2, y2)
-    is the bottom right corner.
-    """
-    outputs = torch.empty_like(inputs)
-    outputs[:, 0] = inputs[:, 0] - inputs[:, 2] / 2
-    outputs[:, 1] = inputs[:, 1] - inputs[:, 3] / 2
-    outputs[:, 2] = inputs[:, 0] + inputs[:, 2] / 2
-    outputs[:, 3] = inputs[:, 1] + inputs[:, 3] / 2
-
-    return outputs
-
-
-def xyxy2xyxyn(inputs: np.ndarray, height: float, width: float) -> np.ndarray:
-    """Converts from [x1, y1, x2, y2] to normalized [x1, y1, x2, y2].
-
-    (x1, y1) is the top left corner and (x2, y2) is the bottom right corner.
-    Normalized coordinates are w.r.t. original image size.
-    """
-    outputs = np.empty_like(inputs)
-    outputs[:, [0, 2]] = inputs[:, [0, 2]] / width
-    outputs[:, [1, 3]] = inputs[:, [1, 3]] / height
-
-    return outputs
